stats estimator support

+++ src/main/scala/io/findify/featury/model/FeatureValue.scala
@@ -7,11 +7,11 @@ object FeatureValue {
   sealed trait ScalarValue[T <: Scalar] extends FeatureValue {
     def value: T
   }
-  case class TextScalarValue(value: Text)                                            extends ScalarValue[Text]
-  case class NumScalarValue(value: Num)                                              extends ScalarValue[Num]
-  case class NumStatsValue(min: Double, max: Double, quantiles: Map[Double, Double]) extends FeatureValue
-  case class PeriodicNumValue(values: List[PeriodicValue])                           extends FeatureValue
-  case class StringFrequencyValue(values: Map[String, Double])                       extends FeatureValue
+  case class TextScalarValue(value: Text)                                         extends ScalarValue[Text]
+  case class NumScalarValue(value: Num)                                           extends ScalarValue[Num]
+  case class NumStatsValue(min: Double, max: Double, quantiles: Map[Int, Double]) extends FeatureValue
+  case class PeriodicNumValue(values: List[PeriodicValue])                        extends FeatureValue
+  case class StringFrequencyValue(values: Map[String, Double])                    extends FeatureValue
   sealed trait BoundedListValue[T <: Scalar] extends FeatureValue {
     def values: List[ListItem[T]]
   }
